disease_model_training/requirements.txt
tensorflow>=2.12.0
datasets>=2.14.0
huggingface-hub>=0.17.0
Pillow>=9.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0

# Data scraping dependencies
pandas>=2.0.0
openpyxl>=3.1.0
pdfplumber>=0.10.0
tabula-py>=2.5.0
